[major] widespread modularization, refactoring utils
