+++ src/scripts/charts/PercentageChart.js
@@ -1,5 +1,5 @@
 import BaseChart from './BaseChart';
-import $ from '../helpers/dom';
+import $ from '../utils/dom';
 
 export default class PercentageChart extends BaseChart {
 	constructor(args) {
