--- src/scripts/helpers/draw.js
@@ -1,25 +1,0 @@
-export function float_2(d) {
-	return parseFloat(d.toFixed(2));
-}
-
-export function make_x_line(height, text_start_at, point, label_class, axis_line_class, x_pos) {
-
-}
-
-export function make_y_line() {}
-
-export function draw_x_line() {}
-
-export function draw_y_line() {}
-
-export function label_x_line() {}
-
-export function label_y_line() {}
-
-export function get_anim_x_line() {}
-
-export function get_anim_y_line() {}
-
-export function draw_rect() {}
-
-export function draw_circle() {}
